fix(theme): wire tauriTheme.current into systemTheme() for live OS changes
Moves systemTheme/applyTheme/resolveTheme into useTheme.svelte.ts so they
read tauriTheme.current ($state updated by onThemeChanged + polling). The
prior systemTheme() in useTheme.ts read only window.__TANDEM_INITIAL_THEME__
(set once at startup) — live OS app-mode flips updated the store but never
reached the DOM because applyTheme resolved against the stale startup value.

Also:
- Fix Ok(Some(tauri::Theme::Dark)) -> Ok(tauri::Theme::Dark) for Tauri 2.10.3
  in both the setup seed match and get_app_theme command
- Add Err arm warn-log to the setup-block theme match
- Replace unload with pagehide + import.meta.hot?.dispose for Chromium/HMR
  reliability; reset pollErrorLogged on poll success for transient recovery
- Rewrite useTauriTheme seeding test to exercise TauriThemeStore constructor
  via vi.resetModules() + vi.doMock() instead of asserting a self-assignment
- Update useTheme.test.ts import to useTheme.svelte.js
